Make sure AtMostOnceBarrier only releases once (or never) (#4)
I realized the implementation of `incr(self, key, amount, maximum, ttl)`
assumes a default value of `0` if the key does not exist. This means,
when the backend (e.g. Redis) loses its contents, we would assume every
`AtMostOnceBarrier` never ran. To fix this, the key is now created with
a value of `-1` (instead of `0`) and we only release if by incrementing
by `1` we stay under the maximum of `0` - which can only be true if the
previous value was `-1` and not with the default of `0`.

I also added some tests to confirm this bug existed and to confirm this
PR fixes it.