More efficient completion callback serialization (#5)
As of now, running a group of workflows would lead to serializing the
`completion_callbacks` separately for each member of the group. With
large workflows, this could be quite slow. Additionally, there was a
long-standing TODO in the middleware to not unserialize the callbacks
since they just have to be serialized again immediately after.

This PR aims to solve both these issues. We never actually need an
unserialized version of the callbacks so we now only pass around the
serialized version. Only the workflow that needs to be run is now
unserialized by the middleware when needed.