Always use barrier (#2)
Previously, we skipped creating a barrier when only one task was
required to complete before continuing the workflow. However, this
sometimes led to duplicate tasks when a task in a `Chain` was
(incorrectly) retried after already having completed previously. To
alleviate this, we now always create an `AtMostOnce` barrier, even if we
only await the completion of a single task.

I also added a bit of documentation about the custom barrier
implementation and a way to switch to the more lenient default `Barrier`
implementation of dramatiq.